Update snow flake generation
Previously, were drawn anywhere. To make it so that flakes could
not draw over each other, logic was added to make sure that they
could only begin drawing if they were not going to collide with
another flake. This lead to an issue where the turtle would get
deadlocked and be unable to draw so logic was added to determine
when that was happening and a flake will be erased.

To go with the logical erasure of flakes, fucntionality to erase
flakes after a certain amount have been drawn was also added to
give the appeareance of snow falling.

The erasing of flakes was encapsulated in its own function as it
is used in multiple places.